requirements.txt
psycopg2==2.9.5
pandas==1.5.3
sqlalchemy==2.0.7
python-dotenv==0.21.1
requests==2.28.2